+++ Ray Tracer/Ray.cpp
@@ -22,7 +22,7 @@ Vector Ray::Direction() const {
     return direction;
 }
 
-Vector Ray::FindPoint(GLfloat t) const {
+Vector Ray::FindPoint(GLdouble t) const {
     return position + (direction * t);
 }
 
@@ -34,11 +34,11 @@ Vector Ray::FindPoint(GLfloat t) const {
 //
 //
 //
-//GLfloat MinimumDistance(const Ray& r1, const Ray& r2) {
+//GLdouble MinimumDistance(const Ray& r1, const Ray& r2) {
 //    Vector v1 = r1.Direction();
 //    Vector v2 = r2.Direction();
-//    GLfloat dot = v1.dot(v2);
-//    GLfloat dotSquared = dot * dot;
+//    GLdouble dot = v1.dot(v2);
+//    GLdouble dotSquared = dot * dot;
 //    
 //    // Lines are parallel
 //    if (dotSquared == 1) {
@@ -47,13 +47,13 @@ Vector Ray::FindPoint(GLfloat t) const {
 //    
 //    Vector originDiff = r2.Position() - r1.Position();
 //    
-//    GLfloat t1 = dot * originDiff.dot(v2) - originDiff.dot(v1);
-//    GLfloat t2 = originDiff.dot(v2) - dotSquared * originDiff.dot(v1);
+//    GLdouble t1 = dot * originDiff.dot(v2) - originDiff.dot(v1);
+//    GLdouble t2 = originDiff.dot(v2) - dotSquared * originDiff.dot(v1);
 //    
 //    Vector point1 = r1.FindPoint(t1);
 //    Vector point2 = r2.FindPoint(t2);
 //    
-//    GLfloat minimumDistance = (point1 - point2).length();
+//    GLdouble minimumDistance = (point1 - point2).length();
 //    
 //    return minimumDistance;
 //}
