+++ Ray Tracer/Sphere.hpp
@@ -18,15 +18,15 @@
 class Sphere: public Shape {
     
 public:
-    Sphere(Vector center, GLfloat radius);
-    Sphere(Vector center, GLfloat radius, Material mat);
+    Sphere(Vector center, GLdouble radius);
+    Sphere(Vector center, GLdouble radius, Material mat);
     
-    GLfloat rayIntersection(const Ray& r) const override;
+    GLdouble rayIntersection(const Ray& r) const override;
     Vector findNormalAtPoint(const Vector& v) const override;
     
 private:
     Vector center;
-    GLfloat radius;
+    GLdouble radius;
 };
 
 #endif /* Sphere_hpp */
